requirements.txt
pyodbc==4.0.26
azure-keyvault==1.1.0
nltk==3.2.5
tika==1.24
beautifulsoup4==4.7.1
exchangelib==3.2.0
pyyaml==5.1
pytz==2018.9
opencensus-ext-azure==1.0.2
azure-eventhub==5.1.0